Manual creation of Module class extension is partially supported.
This model should make a bit more sense to Java developers than the one
we had before, where we had to declare unnecessary instance members
in protected scope in the user defined module class. It should be better
to provide utility class around this current model to make it easier to
build the module than to make developers remember awkward conventions.