Cleaning up after module refactoring.
Keeping the module generator for now, but this could get confusing in
the future. If the user uses module generator to instantiate the user
defined module object, and then tries to built it manually, then it 
might cause some problems. Consider making them compatible, or, remove
one or the other.