Renamed Constructor to Constractable and deprecated it.
Since a GWT bug prevents the model where subtypes of JSObject to implement
common interface (such as Constructable), we are planning to migrate to a
model similar to the one just implemented with Creator interface with a
generator. Planning to use the name Constructor for this, hence the interface
formerly called Constructor was renamed to Constructable (which is more
appropriate anyway) and it was deprecated.