#12 Depends annotation is hidden inside API.
Depends annotation is now used only to refer to the native NG module
names such as "ngRoute" and "ngSanitize", so that the wrappers know
which native modules they depend on. The dependency outside of this is
automatically resolved through the inclusion of the <inherits>
declarations in the gwt.xml, which invokes the EntryPoint of each GWT
module, which then corresponds to the Angular Module. When the
Angular.bootstrap() method is called at the end to start the UI, all
modules previously loaded via EntryPoints are all initiated.