Renamed InjectorFactory to BinderFactory because Injector means something specific in AngularJS.
https://github.com/kyoken74/gwt-angular/issues/32
One of the byproducts of reading is that I know more, and knowing more is
always a big problem. I start seeing problems everywhere. This time, I
realized that the word Injector means something very specific in Angular,
something rather fundamental to the framwork. I've been using that word
to refer to the process responsible for injecting the Angular managed
objects such as services into the "shells" or wrappers written in Java.
This could cause a massive confusion later on if we onboard any developers
who knows Angular to help with the development of this project. Renamed to
BinderFactory, and former BinderFactory was renamed to ScopeBinderFactory
so that we know these processes are used to "bind" JS to Java, either
directory (BinderFactory) or via $scope ScopeBinderFactory.