Adding filtering to set rename-to at build time.
I originally wanted to stick in the ${project.version} in there, so
that the final WAR file in its expanded form would lookg something
like gwt-angular-demo/0.0.19-SNAPSHOT or gwt-angular-demo/0.0.18
but GWT doesn't like to get these names stuck in the rename-to
field, (says it's not a valid name). I don't quit understand why
this restriction exists, because rename-to value is really a run-
time environment concept, and so long as the target platform
supports the directory name, then it should be allowed. In fact,
if I manually rename the directory after the build and the deploy,
the deployed artifact works just fine.