Working on the hash param driven navigation.
This is the first cut, and so far the tab navigation is broken, but I'm
committing the code to mark an intermediate stage. The tab navigation must be
fixed, and, after that, the code should be REFACTORED, because there are
redundant code and resource loading, etc. It's would be embarrasing to release
something like this in the demo code.

https://github.com/kyoken74/gwt-angular/issues/33