Migrated all the partials to ClientBundle. TO REFACTOR LATER
It is working well, but the user has to remember how to assemble a lot
of moving parts to get this working, and the route and the template can
be confusing. We also need to make the templates available from the
HTML. Another code generator might be needed to abstract it and simplify
the interaction between the code and the developer.

https://github.com/kyoken74/gwt-angular/issues/43