Renamed the closure and function method names.
Coming back to the code, the names function and closure were a bit
confusing, because I couldn't tell what they were supposed to do, i.e.
whether they were creation methods or not. Changed the names to call
and exec, which makes a bit more sense to me.

Still working on the directives.

https://github.com/kyoken74/gwt-angular/issues/36