Working on directives.
It's getting there... The inclusion of text as template works. The 
inclusion of data resource (i.e. partial) via templateUrl works. The
restrict seems to be working though not tested, etc, but the one I care
about right now most, the value of the directive attribute, is somehow
not working yet.

https://github.com/kyoken74/gwt-angular/issues/36