Reduced the pre.prettyprint font size to 12px from 13px.
The code sample on the screen on my MacBook 13 inch, Firefox did not
look so good. Maybe after looking at it with other browsers on various
platforms, I will change my mind. For now, 12px is easier for me to read
on the screen, given my screen size and the browser's rendering of the
font.