Cleaned up a lot of the directives. Better scope handling.
I think I finally understood how the directive's scope property is 
supposed to work. When scope is true, it creates a new local scope 
inheriting parent scope. When scope is a JSON object, it creates a new
isolateScope, which does not automatically inherit the parent's scope.

I could not figure out from AngularJS documentation what is scope value
of false or null/undefined is.