ClientResources replaces ClientBundle for code gen
We used to perform code generation directly on ClientBundle. This
presents a problem, because GWT makes heavy use of ClientBundle. To
limit the visibility of the resources intended for injection into
Angular, we extended ClientBundle and created ClientResources. This
extended interface (which only exists for Angular and not for GWT) is
used to GWT Angular's code generation.