Improving error message for Module.resources
I just noticed that when we register a ClientResources descendants that
are not used anywhere in views, we sometimes get benign error hitting
this condition. Improved the error message so that it is clearer that
the direct descendants of ClientBundle do not need to be registered
with Module.resources. Only if the user intends to use the resources
class as a controller in his view that he must extend ClientResources
and register it with Module.resources.