Putting angularjs resource on gwt source path
Addressing the compiler warning

Resource 'bower_components/...' was located via ClassLoader. As a
result changes in that resource will not be reflected in per-file
recompiles. It should be registered via  <source /> or <resource />
entry in your .gwt.xml. In a future version of GWT, we will remove this
fallback and your application will stop compiling