Restored deprecated interface.
The factory method was used a lot previously, and if anyone has been
using GWT Angular 0.0.67 or earlier, there's a good chance they are
using it, too. While we never made any promises to preserve interface
in 0.0.x, we will make an exception here because this could be really
annoying. This method will be removed when we go to 0.1.0.