Introducing a new annotation for code gen
SupportedRootClass annotation will drive the subject type of code
generation. In another words, if we want to create a Factory class for
Fubar type, then we need to annotate the Factory interface with
@SupportedRootClass(Fubar.class) so that at code generation time, we
can scan the classloader for the descendants of Fubar.class and set up
the factory for all the descendants.