#88 Deprecating object instance based factory
The factory method relying on previously instantiated object implies
that, if we were to use these factory methods, we cannot defer the
instantiation of the subject object. This is one of the major problem
that we wish to address. We deprecate this old interface, and will
migrate to a less resource eager model.