#88 Renamed DependenciesFactory to Inspector
The name factory implies that it creates something. Name like
DependenciesFactory can mislead the user, because it can be interpreted
as something that instantiates the objects needed by someone else. In
reality, this class only provides the list of names of components
required by a class. The name DependencyInspector sounds a little
closer to what it actually does, and so it was renamed.