#88 Removed name property from Factory for now
The name of the service used to be stored by the factory object, but
this does not mirror the AngularJS API, which requires the name
parameter to be passed to the .factory() function as the first
parameter. Since I only added it recently, I don't think anyone else is
using it. Removed.