#86 Added comments describing error
Did some debugging, and found that the wrong value is being stored in
ReferenceMapper, but did not have the time to trace it back to the
source of the issue. The binding characteristic for HtmlResource should
be interface, not class, but somehow wrong value is being stored in
ReferenceMapper, causing ClassCastException. This only seem to happen
in SDM when I run it in Eclipse, but when I do standard compilation,
this does not seem to happen.