#86 2.6.1 SDM is slightly more stable than 2.7.0
It still breaks with the error mentioned in the ticket, but at least it
starts up at first. With 2.7.0, the SDM did not start at all, which was
a real pain when making small changes. We really need this fixed before
the framework can be declared useful, because it KILLs productivity.