Making controller init code consistent with others
Now that I started cleaning up Directive and Filter, I noticed that
Controller was not doing the things in the same patter. It was already
similar to them, but I like keeping things consistent, so when I come
up with new and better ideas, it is easier to apply the same solution
throughout the code.